@@ -27,7 +27,8 @@ dependencies:
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
   cupertino_icons: ^1.0.2
-  webview_flutter: ^2.0.10
+  webview_flutter: ^4.7.0
+  webview_flutter_web: ^0.2.2+2
   chopper: ^6.0.2
   connectivity: ^3.0.6
   http: ^0.13.3
